transfere task de arquivo
